% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asOld.r
\name{as.MFCLFrq}
\alias{as.MFCLFrq}
\title{Methods for accessing and creating MFCLFrq and MFCLFrqOld objects}
\usage{
as.MFCLFrq(frq, ...)
}
\description{
Create a function that transforms the new 3 dataframe frq object into the old freq object
Keep this for backward compatability of code
}
