% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrqClasses.r, R/FrqAccessors.r
\name{FrqData}
\alias{FrqData}
\alias{lf_range}
\alias{lf_range<-}
\alias{age_nage}
\alias{age_nage<-}
\alias{cateffpen}
\alias{cateffpen<-}
\alias{lnfrq}
\alias{lnfrq<-}
\alias{wtfrq}
\alias{wtfrq<-}
\alias{range}
\alias{range<-}
\title{FrqData}
\usage{
FrqData()

\S4method{lf_range}{FrqData}(object)

\S4method{lf_range}{FrqData,numeric}(object) <- value

\S4method{age_nage}{FrqData}(object)

\S4method{age_nage}{FrqData,numeric}(object) <- value

\S4method{cateffpen}{FrqData}(object)

\S4method{cateffpen}{FrqData,data.frame}(object) <- value

\S4method{lnfrq}{FrqData}(object)

\S4method{lnfrq}{FrqData,data.frame}(object) <- value

\S4method{wtfrq}{FrqData}(object)

\S4method{wtfrq}{FrqData,data.frame}(object) <- value

\S4method{range}{Frq}(x)

\S4method{range}{Frq,missing,numeric}(x) <- value
}
\description{
Basic constructor for FrqData class
}
