% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readfrq.r
\name{readfrq}
\alias{readfrq}
\title{Frq File read function}
\usage{
readfrq(file)
}
\arguments{
\item{file}{a Character string giving the name and path of the frq file to be read}
}
\value{
An object of class Frq
}
\description{
Extracts all information from the frq file and stores it into various slots
}
\examples{
readFrqData("~/Documents/SKJ/2019/assessment/RefCase/skj.frq")

}
