% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readfrq.r
\name{readFrqData}
\alias{readFrqData}
\title{Frq Size frequency information}
\usage{
readFrqData(file)
}
\arguments{
\item{file}{a Character string giving the name and path of the frq file to be read}
}
\value{
An object of class FrqData
}
\description{
Extracts the catch, size and effort information from the frq file and stores it into three easy to access dataframes
}
\examples{
readFrqData("~/Documents/SKJ/2019/assessment/RefCase/skj.frq")

}
