% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefrq.r
\name{writefrq}
\alias{writefrq}
\title{Function to write a Frq object to file}
\usage{
writefrq(x, ...)
}
\arguments{
\item{x}{object of class}

\item{file}{path and name of file to be written}

\item{append}{whether to add to the end of the file or overwrite}
}
\description{
Function to write a Frq object to file
}
\examples{
writefrq(frq,"skj.frq")

}
